FIX - Issue #32: Renamed test_fun --> test and test --> unittest
